requirements.txt
amqp==5.3.1
asgiref==3.9.1
async-timeout==5.0.1
billiard==4.2.1
boto3==1.40.8
botocore==1.40.8
celery==5.3.6
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.5
Django==5.0.3
django-anymail==10.3
django-celery-beat==2.8.1
django-extensions==4.1
django-filter==24.2
django-redis==6.0.0
django-timezone-field==7.1
djangorestframework==3.15.1
ecdsa==0.19.1
exceptiongroup==1.3.0
idna==3.10
iniconfig==2.1.0
jmespath==1.0.1
kombu==5.5.4
MarkupSafe==3.0.2
model-bakery==1.20.5
numpy==2.2.6
packaging==25.0
pandas==2.2.2
pluggy==1.6.0
prompt_toolkit==3.0.51
psycopg2-binary==2.9.9
pytest==8.2.2
pytest-django==4.8.0
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
python-http-client==3.3.7
pytz==2025.2
redis==5.0.4
requests==2.32.4
s3transfer==0.13.1
sendgrid==6.12.4
six==1.17.0
sqlparse==0.5.3
tomli==2.2.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.1.3
